+++ src/dirtyfields/__init__.py
@@ -2,7 +2,8 @@
 
 from importlib.metadata import version
 
-__all__ = ["DirtyFieldsMixin"]
+__all__ = ["DirtyFieldsMixin", "normalise_value", "raw_compare", "timezone_support_compare"]
 __version__ = version("django-filthyfields")
 
+from dirtyfields.compare import normalise_value, raw_compare, timezone_support_compare
 from dirtyfields.dirtyfields import DirtyFieldsMixin